Change `ticks` to `Int` for tests

+++ src/main/scala/beam/agentsim/agents/ridehail/allocation/RideHailResourceAllocationManager.scala
@@ -61,7 +61,7 @@ abstract class RideHailResourceAllocationManager(private val rideHailManager: Ri
   /*
     This method is called periodically
    */
-  def updateVehicleAllocations(tick: Double, triggerId: Long): Unit = {
+  def updateVehicleAllocations(tick: Int, triggerId: Long): Unit = {
     logger.trace("default implementation updateVehicleAllocations executed")
   }
